move prim/arr/meta to prim/scal/meta
