+++ stan/math/prim/scal/meta/is_constant.hpp
@@ -3,8 +3,11 @@
 
 #include <stan/math/prim/scal/meta/bool_constant.hpp>
 #include <stan/math/prim/scal/meta/conjunction.hpp>
+#include <stan/math/prim/scal/meta/is_vector.hpp>
+#include <stan/math/prim/scal/meta/is_constant.hpp>
+#include <stan/math/prim/scal/meta/require_generics.hpp>
 #include <type_traits>
-
+#include <vector>
 namespace stan {
 
 /** \ingroup type_trait
@@ -31,6 +34,18 @@ struct is_constant : bool_constant<std::is_convertible<T, double>::value> {};
  */
 template <typename... T>
 using is_constant_all = math::conjunction<is_constant<T>...>;
+
+/** \ingroup type_trait
+ * Defines a static member named value and sets it to true
+ * if the type of the elements in the provided std::vector
+ * is constant, false otherwise. This is used in
+ * the is_constant_all metaprogram.
+ * @tparam type of the elements in the std::vector
+ */
+template <typename T>
+struct is_constant<T, require_std_vector_t<T>>
+    : bool_constant<is_constant<typename std::decay_t<T>::value_type>::value> {
+};
 
 }  // namespace stan
 #endif
