+++ stan/math/prim/scal/meta/length.hpp
@@ -2,6 +2,7 @@
 #define STAN_MATH_PRIM_SCAL_META_LENGTH_HPP
 
 #include <cstdlib>
+#include <vector>
 
 namespace stan {
 /** \ingroup type_trait
@@ -11,6 +12,18 @@ namespace stan {
 template <typename T>
 size_t length(const T& /*x*/) {
   return 1U;
+}
+
+/** \ingroup type_trait
+ * Returns the length of the provided std::vector.
+ *
+ * @param x input vector
+ * @tparam T type of the elements in the vector
+ * @return the length of the input vector
+ */
+template <typename T>
+size_t length(const std::vector<T>& x) {
+  return x.size();
 }
 }  // namespace stan
 #endif
