+++ stan/math/prim/scal/meta/scalar_type.hpp
@@ -2,7 +2,10 @@
 #define STAN_MATH_PRIM_SCAL_META_SCALAR_TYPE_HPP
 
 #include <stan/math/prim/scal/meta/value_type.hpp>
+#include <stan/math/prim/scal/meta/is_vector.hpp>
+#include <stan/math/prim/scal/meta/scalar_type.hpp>
 #include <type_traits>
+#include <vector>
 
 namespace stan {
 
@@ -35,6 +38,15 @@ using scalar_type_t = typename scalar_type<T>::type;
 
 template <typename T>
 using scalar_type_decay_t = typename scalar_type<std::decay_t<T>>::type;
+
+/** \ingroup type_trait
+ * Specialization of scalar_type for vector to recursivly return the inner
+ * scalar type.
+ */
+template <typename T>
+struct scalar_type<T, std::enable_if_t<is_std_vector<T>::value>> {
+  using type = scalar_type_t<typename std::decay_t<T>::value_type>;
+};
 
 }  // namespace stan
 #endif
