+++ stan/math/rev/meta/operands_and_partials.hpp
@@ -10,7 +10,7 @@
 #include <stan/math/prim/scal/meta/is_vector_like.hpp>
 #include <stan/math/prim/scal/meta/likely.hpp>
 #include <stan/math/rev/mat/fun/typedefs.hpp>
-#include <stan/math/prim/arr/meta/length.hpp>
+#include <stan/math/prim/scal/meta/length.hpp>
 #include <vector>
 
 namespace stan {
